adds new Celery queue for email tasks
