+++ src/fidesops/ops/service/privacy_request/request_runner_service.py
@@ -1,9 +1,8 @@
 import logging
 import random
 from datetime import datetime, timedelta
-from typing import ContextManager, Dict, List, Optional, Set
+from typing import Dict, List, Optional, Set
 
-from celery import Task
 from celery.utils.log import get_task_logger
 from fideslib.db.session import get_db_session
 from fideslib.models.audit_log import AuditLog, AuditLogAction
@@ -40,7 +39,7 @@ from fidesops.ops.task.graph_task import (
     run_access_request,
     run_erasure,
 )
-from fidesops.ops.tasks import celery_app
+from fidesops.ops.tasks import celery_app, DatabaseTask
 from fidesops.ops.tasks.scheduled.scheduler import scheduler
 from fidesops.ops.util.cache import (
     FidesopsRedis,
@@ -173,19 +172,6 @@ def queue_privacy_request(
         )
 
     return task.task_id
-
-
-class DatabaseTask(Task):  # pylint: disable=W0223
-    _session = None
-
-    @property
-    def session(self) -> ContextManager[Session]:
-        """Creates Session once per process"""
-        if self._session is None:
-            SessionLocal = get_db_session(config)
-            self._session = SessionLocal()
-
-        return self._session
 
 
 @celery_app.task(base=DatabaseTask, bind=True)
